Minor
CryCatScreen: changed capacity get for the meter to return a more static value in case of server-client desync
PurBestBE: changed action validity check to check for the capacity of the tank instead of simulating a fill, which broke when the tank was already full.
StellAccuScreen: Changed all capacity gets to more static values in case of server-client desync.
SupPowScreen: changed capacity get to static value.